requirements.txt
groq>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
asyncio-throttle>=1.0.0
pytest>=7.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pinecone-client>=3.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
jupyter>=1.0.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.20